requirements.txt
astroid==2.4.2
colorama==0.4.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
python-dotenv==0.15.0
six==1.15.0
toml==0.10.2
wrapt==1.12.1


discord~=1.0.1
requests~=2.25.1
matplotlib~=3.3.4
beautifulsoup4~=4.9.3